split out and test
